refacted comments form
