+++ project/src/constants.ts
@@ -105,3 +105,11 @@ export enum Response {
   PENDING,
   ERROR
 }
+
+export const FAVORITES_INPUT_DATA = [
+  { value: '5', id: '5-stars', title: 'perfect' },
+  { value: '4', id: '4-stars', title: 'good' },
+  { value: '3', id: '3-stars', title: 'not bad' },
+  { value: '2', id: '2-stars', title: 'badly' },
+  { value: '1', id: '1-stars', title: 'terribly' },
+];
